User Put Request geupdatet en uitgebreid.
en User Delete Request geupdatet en uitgebreid

+++ src/main/java/nl/novi/eindopdrachtbackend/service/UserService.java
@@ -43,13 +43,15 @@ public class UserService {
         return collection;
     }
 
+    String errormessage = "User not found!";
+
     public UserDto getUser(String username) {
         UserDto userDto = new UserDto();
         Optional<User> user = userRepository.findById(username);
         if (user.isPresent()) {
             userDto = fromUserToDto(user.get());
         } else {
-            throw new UsernameNotFoundException(username);
+            throw new UsernameNotFoundException(errormessage);
         }
         return userDto;
     }
@@ -67,18 +69,34 @@ public class UserService {
     }
 
     public void deleteUser(String username) {
-        userRepository.deleteById(username);
+        Optional<User> optionalUser = userRepository.findById(username);
+        if (optionalUser.isPresent()) {
+            userRepository.deleteById(username);
+
+        } else {
+            throw new UsernameNotFoundException(errormessage);
+        }
     }
 
-    public void updateUser(String username, UserDto newUser) {
-        if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
-        User user = userRepository.findById(username).get();
-        user.setPassword(passwordEncoder.encode(newUser.getPassword()));
-        userRepository.save(user);
+
+    /// HOE OUTDATED USER VERVANGEN VOOR UPDATED USER, IPV TWEE LOSSE USERS?
+    public void updateUser(String username, UserDto updatedUser) {
+        Optional<User> optionalUser = userRepository.findById(username);
+        if (optionalUser.isPresent()) {
+//            User user = optionalUser.get();
+//            user.setPassword(passwordEncoder.encode(updatedUser.getPassword()));
+
+            fromDtoToUser(updatedUser);
+            userRepository.save(fromDtoToUser(updatedUser));
+
+        } else {
+            throw new UsernameNotFoundException(errormessage);
+
+        }
     }
 
     public Set<Authority> getAuthorities(String username) {
-        if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
+        if (!userRepository.existsById(username)) throw new UsernameNotFoundException(errormessage);
         User user = userRepository.findById(username).get();
         UserDto userDto = fromUserToDto(user);
         return userDto.getAuthorities();
@@ -87,7 +105,7 @@ public class UserService {
 
     public void addAuthority(String username, String authority) {
 
-        if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
+        if (!userRepository.existsById(username)) throw new UsernameNotFoundException(errormessage);
         User user = userRepository.findById(username).get();
         user.addAuthority(new Authority(username, authority));
         userRepository.save(user);
@@ -95,7 +113,7 @@ public class UserService {
     //Rollen toevoegen - één rol per keer
 
     public void removeAuthority(String username, String authority) {
-        if (!userRepository.existsById(username)) throw new UsernameNotFoundException(username);
+        if (!userRepository.existsById(username)) throw new UsernameNotFoundException(errormessage);
         User user = userRepository.findById(username).get();
         Authority authorityToRemove = user.getAuthorities().stream().filter((a) -> a.getAuthority().equalsIgnoreCase(authority)).findAny().get();
         user.removeAuthority(authorityToRemove);
